+++ src/debugSession/BrightScriptDebugSession.spec.ts
@@ -10,13 +10,14 @@ import { fileUtils } from '../FileUtils';
 import type { EvaluateContainer, StackFrame, TelnetAdapter } from '../adapters/TelnetAdapter';
 import { PrimativeType } from '../adapters/TelnetAdapter';
 import { defer } from '../util';
-import { HighLevelType, RokuAdapterEvaluateResponse } from '../interfaces';
+import { HighLevelType } from '../interfaces';
 import type { LaunchConfiguration } from '../LaunchConfiguration';
 import type { SinonStub } from 'sinon';
 import { util as bscUtil, standardizePath as s } from 'brighterscript';
 import { DefaultFiles } from 'roku-deploy';
 import type { AddProjectParams, ComponentLibraryConstructorParams } from '../managers/ProjectManager';
 import { ComponentLibraryProject, Project } from '../managers/ProjectManager';
+import { RendezvousTracker } from '../RendezvousTracker';
 
 const sinon = sinonActual.createSandbox();
 const tempDir = s`${__dirname}/../../.tmp`;
@@ -30,6 +31,7 @@ describe('BrightScriptDebugSession', () => {
     let responses = [];
 
     afterEach(() => {
+        fsExtra.emptydirSync(tempDir);
         fsExtra.removeSync(outDir);
         sinon.restore();
     });
@@ -43,7 +45,11 @@ describe('BrightScriptDebugSession', () => {
     let errorSpy: sinon.SinonSpy;
 
     beforeEach(() => {
+        fsExtra.emptydirSync(tempDir);
         sinon.restore();
+
+        //prevent calling DebugSession.shutdown() because that calls process.kill(), which would kill the test session
+        sinon.stub(DebugSession.prototype, 'shutdown').returns(null);
 
         try {
             session = new BrightScriptDebugSession();
@@ -542,6 +548,33 @@ describe('BrightScriptDebugSession', () => {
         });
     });
 
+    describe('initRendezvousTracking', () => {
+        it('clears history when disabled', async () => {
+            const stub = sinon.stub(session, 'sendEvent');
+            const activateStub = sinon.stub(RendezvousTracker.prototype, 'activate');
+            const clearHistoryStub = sinon.stub(RendezvousTracker.prototype, 'clearHistory');
+
+            session['launchConfiguration'].rendezvousTracking = false;
+
+            await session['initRendezvousTracking']();
+            expect(clearHistoryStub.called).to.be.true;
+            expect(activateStub.called).to.be.false;
+        });
+
+        it('activates when not disabled', async () => {
+            const stub = sinon.stub(session, 'sendEvent');
+            const activateStub = sinon.stub(RendezvousTracker.prototype, 'activate');
+            const clearHistoryStub = sinon.stub(RendezvousTracker.prototype, 'clearHistory');
+
+            session['launchConfiguration'].rendezvousTracking = undefined;
+
+            await session['initRendezvousTracking']();
+            expect(clearHistoryStub.called).to.be.true;
+            expect(activateStub.called).to.be.true;
+
+        });
+    });
+
     describe('setBreakPointsRequest', () => {
         let response;
         let args: DebugProtocol.SetBreakpointsArguments;
@@ -646,8 +679,6 @@ describe('BrightScriptDebugSession', () => {
             (session as any).launchConfiguration = {
                 retainStagingFolder: false
             };
-            //stub the super shutdown call so it doesn't kill the test session
-            sinon.stub(DebugSession.prototype, 'shutdown').returns(null);
 
             await session.shutdown();
             expect(stub.callCount).to.equal(2);